Refactor partitioning logic
Prior to this change, the partioning logic had evolved somewhat
organically, become brittle and duplicating logic.
In particular the MSV2Structure class was evolving towards a
god class that performed partioning, as well as extracting
metadata out of subtables.

This commit refactors MSv2Structure to primarily
perform partioning logic and expose the results of this
partitioning.

Responsbility for populating dataset metadata is now offloaded
to dataset factory classes.

Additionally, the FEED table is now used to establish the
canonical set of ANTENNA's that apply to a partition,
rather than the entire ANTENNA table.